Update the shinyTree input when the checkbox is changed separately from the selection
When the checkbox is just another way to select things in the tree, clicking the box will trigger the selection, which then triggers the shiny tree to update. However, if tie_selection is false, a checkbox change alone does not currently update the shinyTree input as it should. This change fixes this behavior and triggers Shiny to detect an input change in the case that checkbox is true and tie_selection is false whenever a checkbox is unchecked or checked or the check all or uncheck all actions are triggered.